script now works!
added code for printing tree to bpp.ctl and maximum number of sequences
for each species. The maximum number of sequences is based on the total
number of individuals for each species which may be greater than the
number of sequences at any particular locus. Presumably bpp handles
this okay?